requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# ── Vector Store ──────────────────────────────────────────────────────────────
chromadb>=0.5.0
faiss-cpu>=1.8.0

# ── Embedding & Re-ranking ────────────────────────────────────────────────────
sentence-transformers>=3.0.0
torch>=2.2.0
numpy>=1.26.0

# ── LLM Providers ────────────────────────────────────────────────────────────
groq>=0.9.0
openai>=1.30.0
httpx>=0.27.0

# ── Data Handling ─────────────────────────────────────────────────────────────
pandas>=2.2.0

# ── Logging ───────────────────────────────────────────────────────────────────
loguru>=0.7.2

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0  # TestClient dependency

# ── Development ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
